requirements.txt
google-generativeai
streamlit
pillow
python-dotenv